an example of bldc machine
I configured this model with rmxprt tool (very useful for a beginner,
like me :smile: ) and used the machine parameters in the paper that is
in the reference list. also I observed the change in the analysis
results by changing some of the variables. I'm planning to arrange the
reference list (that are bookmarked in my browser, downloaded or on
mendeley) and add the papers that I've read about bldc and fractional
slot machines in the following days.